+++ app/src/main/java/com/brandonjamesyoung/levelup/fragments/SettingsPage.kt
@@ -8,6 +8,7 @@ import android.widget.EditText
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
+import androidx.navigation.fragment.NavHostFragment
 import com.brandonjamesyoung.levelup.R
 import com.brandonjamesyoung.levelup.shared.Difficulty as DifficultyCode
 import com.brandonjamesyoung.levelup.data.Difficulty
@@ -16,6 +17,9 @@ import com.brandonjamesyoung.levelup.shared.NavigationHelper
 import com.brandonjamesyoung.levelup.shared.ToastHelper.Companion.showToast
 import com.brandonjamesyoung.levelup.viewmodels.SettingsViewModel
 import dagger.hilt.android.AndroidEntryPoint
+
+const val MAX_ACRONYM_LENGTH = 3
+val ACRONYM_VALIDATION_REGEX = Regex("^[a-zA-Z]+$")
 
 @AndroidEntryPoint
 class SettingsPage : Fragment(R.layout.settings) {
@@ -40,43 +44,134 @@ class SettingsPage : Fragment(R.layout.settings) {
         )
     }
 
-    private fun validateInput() : Boolean {
+    private fun isNumber(str : String) : Boolean {
+        return str.toIntOrNull() != null
+    }
+
+    private fun validateNumInput(
+        editText : EditText,
+        minNumber : Int? = null,
+        maxNumber : Int? = null,
+    ) : Boolean {
+        val textInput = editText.text.toString()
+
+        if (textInput.isBlank()) {
+            editText.error = resources.getString(R.string.not_a_number_error)
+            return false
+        }
+
+        if (!isNumber(textInput)) {
+            editText.error = resources.getString(R.string.not_a_number_error)
+            return false
+        }
+
+        val numInput = textInput.toInt()
+
+        if (minNumber != null && maxNumber != null && numInput !in minNumber..maxNumber) {
+            editText.error = resources.getString(
+                R.string.num_out_of_range_error,
+                minNumber,
+                maxNumber
+            )
+
+            return false
+        }
+
+        if (minNumber != null && maxNumber == null && numInput < minNumber) {
+            editText.error = resources.getString(R.string.num_too_small_error, minNumber)
+            return false
+        }
+
+        if (minNumber == null && maxNumber != null && numInput > maxNumber) {
+            editText.error = resources.getString(R.string.num_too_large_error, maxNumber)
+            return false
+        }
+
         return true
-        // TODO add validation for settings values here
+    }
 
-        // None of the values can be empty
+    private fun validateDifficultySettings() : Boolean {
+        val difficultyInputIds = listOf(
+            R.id.EasyExpInput,
+            R.id.EasyRtInput,
+            R.id.MediumExpInput,
+            R.id.MediumRtInput,
+            R.id.HardExpInput,
+            R.id.HardRtInput,
+            R.id.ExpertExpInput,
+            R.id.ExpertRtInput
+        )
 
-        // Difficulty, experience, level up bonus must be numbers
+        val view = requireView()
+        var difficultySettingsAreValid = true
 
-        // Acronym must exist, be a string and
+        for (id in difficultyInputIds) {
+            val editText = view.findViewById<EditText>(id)
+            val isValid = validateNumInput(editText, 0, 9999)
+            if (!isValid) difficultySettingsAreValid = false
+        }
+
+        return difficultySettingsAreValid
+    }
+
+    private fun validateAcronym() : Boolean {
+        val view = requireView()
+        val editText : EditText = view.findViewById(R.id.PointsAcronymInput)
+        val textInput = editText.text.toString()
+
+        if (textInput.isBlank()) {
+            editText.error = resources.getString(R.string.no_acronym_error)
+            return false
+        }
+
+        val hasOnlyAlphabet = ACRONYM_VALIDATION_REGEX.matches(textInput)
+
+        if (!hasOnlyAlphabet) {
+            editText.error = resources.getString(R.string.only_alpha_allowed_error)
+            return false
+        }
+
+        if (textInput.length > MAX_ACRONYM_LENGTH) {
+            editText.error = resources.getString(R.string.three_characters_limit_error)
+            return false
+        }
+
+        return true
+    }
+
+    private fun validateLvlUpBonus() : Boolean {
+        val view = requireView()
+        val lvlUpBonusInput = view.findViewById<EditText>(R.id.LevelUpBonusInput)
+        return validateNumInput(lvlUpBonusInput, 0, 999)
+    }
+
+    private fun validateInput() : Boolean {
+        val difficultySettingsAreValid = validateDifficultySettings()
+        val acronymIsValid = validateAcronym()
+        val bonusIsValid = validateLvlUpBonus()
+        return difficultySettingsAreValid && acronymIsValid && bonusIsValid
     }
 
     private fun saveSettings() {
         val view = requireView()
 
-        val easyExpInput = view.findViewById<EditText>(R.id.EasyExpInput)
-        val easyRtInput = view.findViewById<EditText>(R.id.EasyRtInput)
-        val mediumExpInput = view.findViewById<EditText>(R.id.MediumExpInput)
-        val mediumRtInput = view.findViewById<EditText>(R.id.MediumRtInput)
-        val hardExpInput = view.findViewById<EditText>(R.id.HardExpInput)
-        val hardRtInput = view.findViewById<EditText>(R.id.HardRtInput)
-        val expertExpInput = view.findViewById<EditText>(R.id.ExpertExpInput)
-        val expertRtInput = view.findViewById<EditText>(R.id.ExpertRtInput)
-
-        val difficultyInputMap = listOf<Triple<DifficultyCode, EditText, EditText>>(
-            Triple(DifficultyCode.EASY, easyExpInput, easyRtInput),
-            Triple(DifficultyCode.MEDIUM, mediumExpInput, mediumRtInput),
-            Triple(DifficultyCode.HARD, hardExpInput, hardRtInput),
-            Triple(DifficultyCode.EXPERT, expertExpInput, expertRtInput)
+        val difficultyInputMap = listOf(
+            Triple(DifficultyCode.EASY, R.id.EasyExpInput, R.id.EasyRtInput),
+            Triple(DifficultyCode.MEDIUM, R.id.MediumExpInput, R.id.MediumRtInput),
+            Triple(DifficultyCode.HARD, R.id.HardExpInput, R.id.HardRtInput),
+            Triple(DifficultyCode.EXPERT, R.id.ExpertExpInput, R.id.ExpertRtInput)
         )
 
         val newDifficulties = mutableListOf<Difficulty>()
 
         for (triple in difficultyInputMap) {
+            val expInput : EditText = view.findViewById(triple.second)
+            val rtInput : EditText = view.findViewById(triple.third)
+
             val newDifficulty = Difficulty(
                 code = triple.first,
-                expReward = Integer.parseInt(triple.second.text.toString()),
-                rtReward = Integer.parseInt(triple.third.text.toString())
+                expReward = Integer.parseInt(expInput.text.toString()),
+                rtReward = Integer.parseInt(rtInput.text.toString())
             )
 
             newDifficulties.add(newDifficulty)
@@ -100,6 +195,10 @@ class SettingsPage : Fragment(R.layout.settings) {
         confirmButton.setOnClickListener{
             if (validateInput()) {
                 saveSettings()
+                // TODO Settings should send player back to the page they came from
+                //  Default to questList page if no prev page is found
+                NavHostFragment.findNavController(this)
+                    .navigate(R.id.action_settings_to_questList)
             }
         }
     }